вынес классы исключений в отдельные файлы и добавил больше тестовых массивов
