+++ src/Lesson2/Exceptions/MainClass.java
@@ -1,0 +1,92 @@
+package Lesson2.Exceptions;
+
+/*  1. Напишите метод, на вход которого подаётся двумерный строковый массив размером 4х4, при подаче массива другого размера необходимо бросить исключение MyArraySizeException.
+    2. Далее метод должен пройтись по всем элементам массива, преобразовать в int, и просуммировать. Если в каком-то элементе массива преобразование не удалось (например, в ячейке лежит символ или текст вместо числа), должно быть брошено исключение MyArrayDataException, с детализацией в какой именно ячейке лежат неверные данные.
+    3. В методе main() вызвать полученный метод, обработать возможные исключения MySizeArrayException и MyArrayDataException, и вывести результат расчета.
+*/
+
+import static java.lang.Integer.valueOf;
+
+public class MainClass {
+
+    public static void main(String[] args) {
+
+        String[][] stringArray0 = new String[][] {
+                {"5", "g", "3", "4"},
+                {"1", "2", "3", "4"},
+                {"1", "2", "5", "6"},
+                {"1", "2", "4", "8"}
+        };
+        String[][] stringArray1 = new String[][] {
+                {"5", "6", "3", "4"},
+                {"1", "2", "3", "4"},
+                {"1", "2", "5", "6"},
+                {"1", "2", "4", "8"}
+        };
+        String[][] stringArray2 = new String[][] {
+                {"5", "g", "3"},
+                {"1", "2", "3", "4"},
+                {"1", "2", "5", "6"},
+                {"1", "2", "4", "8"}
+        };
+        String[][] stringArray3 = new String[][] {
+                {"1", "2", "3", "4"},
+                {"1", "2", "5", "6"},
+                {"1", "2", "4", "8"}
+        };
+
+        try {
+            System.out.println("Сумма элементов массива равна " + transformAndSum(stringArray0));
+        } catch (MyArraySizeException | MyArrayDataException e) {
+            System.out.println(e + " " + e.getMessage());
+        }
+        try {
+            System.out.println("Сумма элементов массива равна " + transformAndSum(stringArray1));
+        } catch (MyArraySizeException | MyArrayDataException e) {
+            System.out.println(e + " " + e.getMessage());
+        }
+        try {
+            System.out.println("Сумма элементов массива равна " + transformAndSum(stringArray2));
+        } catch (MyArraySizeException | MyArrayDataException e) {
+            System.out.println(e + " " + e.getMessage());
+        }
+        try {
+            System.out.println("Сумма элементов массива равна " + transformAndSum(stringArray3));
+        } catch (MyArraySizeException | MyArrayDataException e) {
+            System.out.println(e + " " + e.getMessage());
+        }
+    }
+
+    private static int transformAndSum (String[][] in) throws MyArraySizeException, MyArrayDataException {
+        int arrDim = 4;
+        int sum = 0;
+
+        // Проверяем, что у массива 4 основных измерения
+        if (in.length != 4) {
+            throw new MyArraySizeException(String.format("Размерность массива должна быть %dх%d.", arrDim, arrDim));
+        }
+        // Т.к. размерность вложенных массивов может быть не всегда одинакова, то проверяем, что в каждом из 4-х основных массивов находится массив ровно из 4 элементов
+        for (int i = 0; i < in.length; i++) {
+            if (in[i].length != 4) {
+                throw new MyArraySizeException(String.format("Размерность массива должна быть %dх%d.", arrDim, arrDim));
+            }
+        }
+
+        for (int i = 0; i < in.length; i++) {
+            for (int j = 0; j < in[i].length; j++) {
+                try {
+                    sum += valueOf(in[i][j]);
+                } catch (NumberFormatException e) {
+                    throw new MyArrayDataException(String.format("В позиции [%d][%d] исходного массива находится не целое число %s.", i, j, in[i][j]));
+                }
+            }
+        }
+
+        return sum;
+    }
+
+}
+
+
+
+
