+++ src/Lesson2/Exceptions/MyArrayDataException.java
@@ -1,0 +1,12 @@
+package Lesson2.Exceptions;
+
+class MyArrayDataException extends RuntimeException {
+    MyArrayDataException(String msg) {
+        super(msg);
+    }
+
+    @Override
+    public String toString() {
+        return "Невозможно найти сумму элементов массива.";
+    }
+}
